Add self.get_updates(callback, period).
I need both series and episode updates, so I wrote a new method that
returns all update info at once: get_updates().  It avoids a bug in
get_updated_shows() (which returns a bunch of None elements from the
nested Series data in Episode and Banner records).  It also makes the
update time available (via the "time" attr).  Finally, it changes the
URL used for all get_update*() functions to fetch the zip-file version,
for reduced transfer sizes.  As for speed, when I processed a week's
worth of updates, the runtime for my script went down to 0.5 seconds
from 1 second (compared to a non-expat functiom I was working on that
returned both episode and series data via the ET method that the other
functions use).