Add "atleast = 1" arg to get_show_and_episodes().
In my use of get_show_and_episodes() I only need newer episodes to be
returned, so I added an arg to get_show_and_episodes() that allows a
DB ID to be specified (it defaults to 1 for backward compatibility).
Specifying this value will stop the routine from returning any episode
objects with a DB ID lower than the atleast request (avoiding a bunch
of episode instantiations for older episodes).