Switch to the thedvdb with various improvements.
This version of Next Aired grabs its data from TheTVDB.com.  This
provides quicker updates, both because we make use of the the update zip
files and because the API seems much faster than the one being used at
tvrage.  This also provides more future-episode information, which we
use to create a better 7-day calendar view.

The settings have changed to specify how often (in hours) background
updates occur (the default is every 12 hours).  There is also a setting
for how many minutes to wait between interactive updates (by default we
update when when entering the guide if at least 10 minutes have gone by
since the last update).  Since updates usually take a few seconds (and
are pretty small) the frequent updates should be fine.

The GUI now includes separate episode info for each show that appears
multiple times in the upcoming week (rather than showing the same next-
episode info for each one).

The WINDOW properties that are set are the same as before.  Only the
Classification is not available (I was able to get the Country info by
looking up the show's TV station).

The data saved is in a new format in a file named next.aired.db (old
filenames from prior versions are cleaned out).  Some simple locking
has been put in place in order to prevent the gui process from starting
a second update pass when the background process is doing an update (the
gui proc will display the background proc's progress while waiting for
it to complete).

We require a tweaked version of thetvdbapi.py file, which was committed
prior to this change.  When the shared version is updated to have the
new features, we can remove our local copy.