Several more update-related improvements.
If updates are failing, stop trying to fetch more data and just tweak
the data that we have.  If the user enters the Guide and it is not yet
time for a user-update, we don't lock or wait for the background proc or
do any update-related fetching at all (even if some data is still marked
as needing an update).  Improved the update-cancel code so that we only
stop fetching update data (we don't abort tweaking known data, which had
resulted in some data being lost until the next update).  The code now
also uses a different socket timeout for background updating (longer)
compared to user updating.