Daily data tweaks & improved failure backoff.
Run update_data() at startup and when the day changes, even if it is not
yet time to grab new data.  This ensures that the episode lists get aged
and that the window properties get set.  The failure handling was also
improved to retry sooner, and back-off if the failures continue.