Add a new Today-oriented TV Guide style.
Added a few extra options that allows the user to include more upcoming
data in their TV Guide, now in 2 styles: 7-day (traditional) and Today-
start (starts with Today or Yesterday).  The 7-day style guide can be
told to include next week's episodes in each day-of-the-week row (they
are added after the current-week's shows).  The Today-start style will
display each day's shows in a separate row, with the days starting with
Today or (if enabled) Yesterday.  Skinners will need to include a new
xml skin to handle the new today-oriented display type.  I've created
an initial release skin, but it could use some refinement.