Added service to start background update on login.
The addon.xml now lists a service that is started on login.  This will
remove the need for a skin to start up the background updating (and we
have detection built-in to ensure that we don't launch it twice -- the
skin's version will auto-exit right after it is started up).